Move machine event configuration to the .machine file for <MyTestimonials>
