+++ src/features/testimonials/MyTestimonials.machine.js
@@ -1,7 +1,9 @@
 import { createMachine, assign } from "xstate";
+import { useMachine } from "@xstate/react";
+import { useMemo } from "react";
 
 /** @xstate-layout N4IgpgJg5mDOIC5QFkCeACALnTBLAtgPYB2uAhgDawB0AMoWRLsVAMQQljXMBuhA1lwBmYTAGMAFgFowPMACdUmCcyiJQAB0KxceEupAAPRAEYADAHZqADgCcJ2wGZrZ6yYCsj9xYBsAGhBUREcTE2oAJgsAFh8zcJiLDwtfAF8UgLQsHAIScipqACUwRlRqAEkICjBWAGEJMhYwLHkyMX50WDAqsT1iAy0dXoNjBHMLd2oo20jwxzMnM0cLAKDRnysouJMor28oixcLNIyMbFg8IlJKGiKS8srquoaYdEJ5CAVVfu1dXH0kIyIdz2ahecIuEzrNybKIrUzrSZbHY+cHhDzuY4gTJnC65a6FYoQUoVKq1eqNdAaMgwb6DP59AEjCxo6i2CFmEx2dzcjxwhAhML2ExeHy2bwo6zhTHY7KXPI3QnEh5k55NMCGBRiXCdDrFeSSWm-f6gEbWaLUaLzHxeCxs+awwLBUKs0JeKLbbzucIY9JY06yvH5W5EugMJgsdicbjEPiCagicTSHE5K5UQ1DRmmYVRCKQnbWSXWKY+ax8-a2ajW6buFzc6zrNK+4iED7wAEy84p+WhxhfAEDI0Mk2IKLhPm2Mysif2cK2Nkmb2j6X+zty-HBpVVdP04aIayOCvhb2WCz28KLfyO0a2HMFmvTZKjuyc5dZVeBhV3ei9ljb42AhBRWsSY5n3KICwXMwfEvVYom5ahtjMbw2W8bNrFfZM1yDRU-yHAD3ARMEIShRCHVWG9JlCSJ3A5ZwCK8DCA1TNtNB+DNh1GUsr05agzD4swokE9w4MWRIfTSIA */
-export default createMachine(
+const machine = createMachine(
   {
     context: {
       parameters: {
@@ -138,4 +140,21 @@ function populateTrackData({ tracks, testimonials }) {
 
 function tracksMap(json) {
   return new Map(json.tracks.map((track) => [track.slug, track]));
+}
+
+export default function useTestimonalsMachine() {
+  const [state, send] = useMachine(machine);
+
+  const events = useMemo(
+    () => ({
+      onPageChange: (page) => send("Change page", { page }),
+      onTrackChange: (track) => send("Change track selection", { track }),
+      onExerciseChange: (exercise) =>
+        send("Change exercise search", { exercise }),
+      onOrderChange: (order) => send("Change ordering", { order }),
+    }),
+    [send]
+  );
+
+  return [state, events];
 }
